creates request and return DTO

+++ src/main/java/com/andresen/macrocalculatorbackend/UserProfileRepository.java
@@ -2,5 +2,5 @@ package com.andresen.macrocalculatorbackend;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserProfileRepository extends JpaRepository<UserProfile, Integer> {
+public interface UserProfileRepository extends JpaRepository<UserProfile, Long> {
 }
